move gl context wrapper down into renderlib
