+++ agave_app/renderer.h
@@ -28,28 +28,6 @@ class IRenderWindow;
 class RenderSettings;
 class Scene;
 
-class RendererGLContext
-{
-public:
-  RendererGLContext();
-  ~RendererGLContext();
-  void configure(QOpenGLContext* glContext = nullptr);
-  void init();
-  void destroy();
-
-  void makeCurrent();
-  void doneCurrent();
-
-private:
-  bool m_ownGLContext;
-#if HAS_EGL
-  HeadlessGLContext* m_glContext;
-#else
-  QOpenGLContext* m_glContext;
-  QOffscreenSurface* m_surface;
-#endif
-};
-
 class Renderer
   : public QThread
   , public RendererCommandInterface
